GNUmakefile: fix install error "argument list too long"
When packaging uutils for Conda-Forge, the install of completion scripts
fails with the following error.

  make: /bin/sh: Argument list too long

The completion scripts are installed with a `foreach` loop that expands
into a single long command. That command hits the maximum argument
length on the build platform used by Conda-Forge.

To fix, we change the Make command to split each install into a seperate
command. This is done by embedding a newline in the command. See the
following for an example of how this works.

https://www.extrema.is/blog/2021/12/17/makefile-foreach-commands